% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_heatmap.R
\name{annotate_heatmap}
\alias{annotate_heatmap}
\title{Add row annotations to a heatmap based on columns from the `row_data` entry
of a `heatmap_data` object}
\usage{
annotate_heatmap(x, row_annotations = NULL, row_dots = NULL,
  top_annotations = NULL, top_dots = NULL)
}
\arguments{
\item{x}{A heatmap_data object.}

\item{row_annotations}{Which columns of the `row_data` entry in `x`
should be used when plotting annotations alongside the rows of the
heatmap?}

\item{row_dots, top_dots}{Additional arguments to be passed to
`ComplexHeatmap::HeatmapAnnotation` for the next-to-the-heatmap
(row_dots) and above-the-heatmap (top_dots) annotations. For example,
`show_legend`, `col`. Check the docs for `HeatmapAnnotation` for parameter
names.}

\item{top_annotations}{Which columns of the `column_data` entry in
`x` should be used when plotting annotations above the heatmap?}
}
\description{
Add row annotations to a heatmap based on columns from the `row_data` entry
of a `heatmap_data` object
}
