% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_heatmap.R
\name{annotate_rows}
\alias{annotate_rows}
\title{annotate_rows}
\usage{
annotate_rows(x, annotations = NULL, ..., replace = FALSE)
}
\arguments{
\item{x}{A heatmap_data object.}

\item{annotations}{Which columns of the `column_data` entry in
`x` should be used when plotting annotations next to the heatmap?}

\item{...}{Additional arguments to be passed to
`ComplexHeatmap::HeatmapAnnotation` for the beside-the-heatmap
annotations. For example, `show_legend`, `col`. Check the docs for
`HeatmapAnnotation` for parameter names.}

\item{replace}{If TRUE, then any existing row-annotation
arguments stored in the heatmap_data will be discarded and replaced with
the provided arguments. If FALSE, the provided arguments will append-to or
overwrite the existing values.}
}
\description{
annotate_rows
}
