% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_colours.R
\name{boolean_colour}
\alias{boolean_colour}
\title{`boolean_colour` return a vector that is formatted for colouring a boolean
variable within a ComplexHeatmap row or column annotation}
\usage{
boolean_colour(true_colour, false_colour = "grey")
}
\arguments{
\item{true_colour}{a colour (string, rgb etc) to indicate the TRUE
value. Required}

\item{false_colour}{a colour (string, rgb etc) to indicate the FALSE
value. Default: grey}
}
\description{
`boolean_colour` return a vector that is formatted for colouring a boolean
variable within a ComplexHeatmap row or column annotation
}
