% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_colours.R
\name{define_annotation_colours}
\alias{define_annotation_colours}
\title{define_annotation_colours: make a named list of colour palettes (discrete
valued colour palettes or functions that return interpolated colours)}
\usage{
define_annotation_colours(boolean_var_names, custom_colour_list = list(),
  true_colour = "black", false_colour = "grey")
}
\arguments{
\item{boolean_var_names}{a character vector. each of the entries
corresponds to a boolean variable that will be presented in the heatmap.
Boolean state will be indicated using true_colour and false_colour. If you
want more control for specific boolean vars, use `boolean_colour` and
`custom_colour_list`}

\item{custom_colour_list}{a named list. Each entry corresponds to
an annotation variable for a heatmaps rows or columns. The names should
match an annotation variable. The values may be functions or named vectors
mapping observable values for the annotation variable to a colour.}

\item{true_colour, false_colour}{default colours for any TRUE/FALSE
variables that are present in `boolean_var_names`}
}
\description{
There should be no overlap between the names defined in `boolean_var_names`
and the `names` of the list `custom_colour_list` since any given variable
can have it's colour-mapping set at most one time.
}
