% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_heatmap.R
\name{format_heatmap}
\alias{format_heatmap}
\title{format_heatmap}
\usage{
format_heatmap(x)
}
\arguments{
\item{x}{A heatmap_data object. As returned by
`setup_heatmap`.}
}
\description{
format_heatmap
}
