% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{? should be plot.heatmap_data method}
\usage{
plot_heatmap(x)
}
\arguments{
\item{x}{A `heatmap_data` object. As returned by
`setup_heatmap` or `format_heatmap`.}
}
\description{
? should be plot.heatmap_data method
}
