% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_heatmap.R
\name{setup_heatmap}
\alias{setup_heatmap}
\title{setup_heatmap: defines the data-structures for use in constructing heatmaps}
\usage{
setup_heatmap(x)
}
\arguments{
\item{x}{A named-list or poly_frame of dataframes /
tibbles. There should, at least, be a data-frame called `bodydata`. This
will be converted to a matrix when plotting.}
}
\description{
setup_heatmap: defines the data-structures for use in constructing heatmaps
}
