% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_heatmap_formatting.R
\name{setup_heatmap_formatting}
\alias{setup_heatmap_formatting}
\title{setup_heatmap_formatting}
\usage{
setup_heatmap_formatting(x)
}
\arguments{
\item{x}{A heatmap_data object. As returned by
`setup_heatmap`.}
}
\description{
setup_heatmap_formatting
}
