% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_heatmap.R
\name{format_heatmap}
\alias{format_heatmap}
\title{format_heatmap}
\usage{
format_heatmap(x, ...)
}
\arguments{
\item{x}{A heatmap_data object. As returned by
`setup_heatmap`.}

\item{...}{Any user-specific formatting flags to be passed
to Heatmap(). An exception is `split`: if this flag is set it should
define a vector of column-names from the `row_data` part of `x`, these
columns are then extracted and passed to Heatmap().}
}
\description{
format_heatmap
}
