% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_heatmap.R
\name{setup_heatmap}
\alias{setup_heatmap}
\title{setup_heatmap: defines the data-structures for use in constructing heatmaps}
\usage{
setup_heatmap(x, row_index = "feature_id", column_index = "sample_id",
  value_index = "fitted_value")
}
\arguments{
\item{x}{A poly_frame or list of named data-frames. This
must contain a data-frame called body_data which defines the data used in
the body of the heatmap (by two index columns and a value column).
Data for use in column- or row-annoations should be passed in as the
`row_data` and `column_data` entries of the `list` `x`.}

\item{row_index, column_index, value_index}{In the data-frames
provided, which column contains the data for the rows / columns / body of
the heatmap?}
}
\description{
setup_heatmap: defines the data-structures for use in constructing heatmaps
}
