% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_heatmap.R
\name{setup_heatmap}
\alias{setup_heatmap}
\title{setup_heatmap: defines the data-structures for use in constructing heatmaps}
\usage{
setup_heatmap(x)
}
\arguments{
\item{x}{A poly_frame or list of named data-frames. This
must contain a data-frame called body_data which defines the data used in
the body of the heatmap (by two index columns and a value column).}
}
\description{
setup_heatmap: defines the data-structures for use in constructing heatmaps
}
