% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_heatmap.R
\name{setup_heatmap}
\alias{setup_heatmap}
\title{setup_heatmap: defines the data-structures for use in constructing heatmaps}
\usage{
setup_heatmap()
}
\description{
setup_heatmap: defines the data-structures for use in constructing heatmaps
}
